Conflicts:
	fastrpcgen/src/com/eprosima/fastrpc/idl/templates/TopicsPluginSourceByInterface.stg
	fastrpcgen/src/com/eprosima/fastrpc/idl/templates/TopicsPluginSourceByOperation.stg
	include/fastrpc/protocols/dds/MessageHeader.h
	include/fastrpc/transports/dds/ProxyTransport.h
	include/fastrpc/transports/dds/ServerTransport.h
	include/fastrpc/transports/dds/TCPProxyTransport.h
	include/fastrpc/transports/dds/TCPServerTransport.h
	include/fastrpc/transports/dds/UDPProxyTransport.h
	include/fastrpc/transports/dds/UDPServerTransport.h
	include/fastrpc/transports/dds/components/ServerProcedureEndpoint.h
	src/protocols/dds/MessageHeader.cpp
	src/protocols/dds/MessageHeaderPlugin.cpp
	src/transports/dds/DDSTCPProxyTransport.cpp
	src/transports/dds/DDSTCPServerTransport.cpp
	src/transports/dds/DDSUDPProxyTransport.cpp
	src/transports/dds/DDSUDPServerTransport.cpp
	src/transports/dds/ProxyTransport.cpp
	src/transports/dds/ServerTransport.cpp
	src/transports/dds/components/ProxyProcedureEndpoint.cpp
	src/transports/dds/components/ServerProcedureEndpoint.cpp
	thirdparty/fastcdr
	thirdparty/idl
	utils/pcTests/dds/MultithreadTest/MultithreadTestClientExample-i86Win32VS2013.vcxproj
	utils/pcTests/dds/MultithreadTest/MultithreadTestClientExample-x64Win64VS2013.vcxproj
	utils/pcTests/rti/MultithreadTest/MultithreadTestClientExample-i86Win32VS2013.vcxproj
	utils/pcTests/rti/MultithreadTest/MultithreadTestClientExample-x64Win64VS2013.vcxproj

+++ examples/C++/DDS/c++11Types/CalculatorTCP/CalculatorTopicsPlugin.cxx
@@ -337,7 +337,7 @@ CalculatorRequestPlugin::get_serialized_sample_max_size(
 
     }
 
-    current_alignment += CalculatorRequest::getMaxCdrSerializedSize(current_alignment);
+    current_alignment += (unsigned int)CalculatorRequest::getMaxCdrSerializedSize(current_alignment);
 
     if (include_encapsulation) {
         current_alignment += encapsulation_size;
@@ -395,7 +395,7 @@ CalculatorRequestPlugin::get_serialized_sample_size(
     return current_alignment - initial_alignment;
     */
 
-    return sample->getSerializedSize(current_alignment);
+    return (unsigned int)sample->getSerializedSize(current_alignment);
 }
 
 unsigned int 
@@ -950,7 +950,7 @@ CalculatorReplyPlugin::get_serialized_sample_max_size(
     }
 
 
-    current_alignment += CalculatorReply::getMaxCdrSerializedSize(current_alignment);
+    current_alignment += (unsigned int)CalculatorReply::getMaxCdrSerializedSize(current_alignment);
 
     if (include_encapsulation) {
         current_alignment += encapsulation_size;
@@ -1008,7 +1008,7 @@ CalculatorReplyPlugin::get_serialized_sample_size(
     return current_alignment - initial_alignment;
     */
 
-    return sample->getSerializedSize(current_alignment);
+    return (unsigned int)sample->getSerializedSize(current_alignment);
 }
 
 unsigned int
